Fix all warnings for -Wall -Wextra -Wno-deprecated-declarations -Werror clean build
- Add (void) casts for unused GTK callback parameters (38 files)
- Fix calloc argument order: calloc(n, sizeof(T)) not calloc(sizeof(T), n)
- Add /* fall through */ comments for intentional switch fall-through
- Fix sign-compare warnings with explicit casts
- Suppress cast-function-type in GTypeInfo with pragma
- Use named initializers for GTypeInfo with value_table field

+++ src/log.c
@@ -42,7 +42,7 @@ void log_func(const gchar *file, gint line, LOG_LEVEL level, const gchar *messag
 	vsnprintf(str, 1024, message, ap);
 	va_end(ap);
 
-	if(level <= ebview_log_level) {
+	if((gint)level <= ebview_log_level) {
 		char *prefix;
 
 		switch(level){
